# Description

A high number of string allocations were found to come from `DefaultUnleash.DetermineIsEnabledAndStrategy`:
![image](https://github.com/Unleash/unleash-client-dotnet/assets/132582942/aababba4-9630-4ef7-99e7-b5564ffda3ae)

I believe this is due to the logging that's done if the toggle isn't available:
![image](https://github.com/Unleash/unleash-client-dotnet/assets/132582942/82b07f9b-3f6e-4114-a7ab-cebf7ca24552)

This will always allocate a string even if logging is disabled (or log level Warning is disabled). We should instead be passing a `Func<string>` which is a smaller allocation that will only get evaluated if logging is on.

Also, in looking at this I noticed there was inconsistant checking if a log level was enabled in `LogExtensions` so I will clean that up as well.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Using the following benchmark (note IsEnabled2 has my changes):
```
[MemoryDiagnoser]
public class Benchmark
{
    readonly UnleashContext context = new();
    readonly DefaultUnleash client = new(new UnleashSettings());

    public Benchmark()
    {
    }

    [Benchmark(Baseline = true)]
    public bool Current() => client.IsEnabled("Moo", context);

    [Benchmark]
    public bool New() => client.IsEnabled2("Moo", context);
}
```

The following results were obtained:
| Method     | Mean     | Error   | StdDev  | Ratio | Gen0   | Allocated | Alloc Ratio |
|----------- |---------:|--------:|--------:|------:|-------:|----------:|------------:|
| Current    | 179.3 ns | 1.80 ns | 1.69 ns |  1.00 | 0.0408 |     512 B |        1.00 |
| New | 105.2 ns | 1.00 ns | 0.94 ns |  0.58 | 0.0254 |     320 B |        0.63 |