@nathanmascitelli 2023-12-21
@daveleek there are three parts of the changes to `LibLog` that would have an impact on performance. The first is updating the methods to always check if the log level is enabled before trying to log.  So if you look at this method:

```
public static void Error(this ILog logger, Func<string> messageFunc)
{
    GuardAgainstNullLogger(logger);
    logger.Log(LogLevel.Error, WrapLogInternal(messageFunc));
}
```
While it checks to see if the logger is null it does not check to see if logging is enabled for `LogLevel.Error` before logging. This does have a performance impact as the underlying logger implementation may or may not actually log without this check.

The second change is the removal of wrapping the input message functions in `WrapLogInternal`, as seen in the example above. I was not sure what exactly this was trying to accomplish to be honest, we have a similar logging interface in our application and just pass the message functions around without wrapping.

The final change is the removal of the methods that accept a `string` instead of `Func<string>`. Allocating a string will always use more memory then allocating a function if that function is not invoked. Given various log levels may not be invoked.

Other than those things everhting else is athestic. Looking at the compiler directives I'm guessing this is public when the SDK is published? I missed that, I can understand that changing the API interface would be a problem.

How about I revert the interface changes so that the methods that were removed are present if `LIBLOG_PUBLIC` is true? That way internally to the SDK we can always use functions but if anyone else was using the string methods they would be unaffected?